+++ tooling/api/src/window.ts
@@ -18,6 +18,9 @@
  *         "center": true,
  *         "requestUserAttention": true,
  *         "setResizable": true,
+ *         "setMaximizable": true,
+ *         "setMinimizable": true,
+ *         "setClosable": true,
  *         "setTitle": true,
  *         "maximize": true,
  *         "unmaximize": true,
@@ -323,7 +326,10 @@ class WebviewWindowHandle {
   }
 
   /**
-   * Listen to an event emitted by the backend that is tied to the webview window.
+   * Listen to an event emitted by the backend or webview.
+   * The event must either be a global event or an event targetting this window.
+   *
+   * See {@link WebviewWindow.emit | `emit`} for more information.
    *
    * @example
    * ```typescript
@@ -336,10 +342,11 @@ class WebviewWindowHandle {
    * unlisten();
    * ```
    *
+   * Note that removing the listener is required if your listener goes out of scope e.g. the component is unmounted.
+   *
    * @param event Event name. Must include only alphanumeric characters, `-`, `/`, `:` and `_`.
    * @param handler Event handler.
    * @returns A promise resolving to a function to unlisten to the event.
-   * Note that removing the listener is required if your listener goes out of scope e.g. the component is unmounted.
    */
   async listen<T>(
     event: EventName,
@@ -356,7 +363,8 @@ class WebviewWindowHandle {
   }
 
   /**
-   * Listen to an one-off event emitted by the backend that is tied to the webview window.
+   * Listen to an one-off event.
+   * See {@link WebviewWindow.listen | `listen`} for more information.
    *
    * @example
    * ```typescript
@@ -369,10 +377,11 @@ class WebviewWindowHandle {
    * unlisten();
    * ```
    *
+   * Note that removing the listener is required if your listener goes out of scope e.g. the component is unmounted.
+   *
    * @param event Event name. Must include only alphanumeric characters, `-`, `/`, `:` and `_`.
    * @param handler Event handler.
    * @returns A promise resolving to a function to unlisten to the event.
-   * Note that removing the listener is required if your listener goes out of scope e.g. the component is unmounted.
    */
   async once<T>(event: string, handler: EventCallback<T>): Promise<UnlistenFn> {
     if (this._handleTauriEvent(event, handler)) {
@@ -386,11 +395,33 @@ class WebviewWindowHandle {
   }
 
   /**
-   * Emits an event to the backend, tied to the webview window.
+   * Emits an event to the backend and all Tauri windows.
+   * The event will have this window's {@link WebviewWindow.label | label} as {@link Event.windowLabel | source window label}.
+   *
    * @example
    * ```typescript
    * import { appWindow } from '@tauri-apps/api/window';
    * await appWindow.emit('window-loaded', { loggedIn: true, token: 'authToken' });
+   * ```
+   *
+   * This function can also be used to communicate between windows:
+   * ```typescript
+   * import { appWindow } from '@tauri-apps/api/window';
+   * await appWindow.listen('sync-data', (event) => { });
+   *
+   * // on another window...
+   * import { WebviewWindow } from '@tauri-apps/api/window';
+   * const otherWindow = WebviewWindow.getByLabel('other')
+   * await otherWindow.emit('sync-data');
+   * ```
+   *
+   * Global listeners are also triggered:
+   * ```typescript
+   * import { appWindow } from '@tauri-apps/api/window';
+   * import { listen } from '@tauri-apps/api/event';
+   * await listen('ping', (event) => { });
+   *
+   * await appWindow.emit('ping');
    * ```
    *
    * @param event Event name. Must include only alphanumeric characters, `-`, `/`, `:` and `_`.
@@ -634,6 +665,33 @@ class WindowManager extends WebviewWindowHandle {
   }
 
   /**
+   * Gets the window's current focus state.
+   * @example
+   * ```typescript
+   * import { appWindow } from '@tauri-apps/api/window';
+   * const focused = await appWindow.isFocused();
+   * ```
+   *
+   * @returns Whether the window is focused or not.
+   *
+   * @since 1.4
+   * */
+  async isFocused(): Promise<boolean> {
+    return invokeTauriCommand({
+      __tauriModule: 'Window',
+      message: {
+        cmd: 'manage',
+        data: {
+          label: this.label,
+          cmd: {
+            type: 'isFocused'
+          }
+        }
+      }
+    })
+  }
+
+  /**
    * Gets the window's current decorated state.
    * @example
    * ```typescript
@@ -677,6 +735,96 @@ class WindowManager extends WebviewWindowHandle {
           label: this.label,
           cmd: {
             type: 'isResizable'
+          }
+        }
+      }
+    })
+  }
+
+  /**
+   * Gets the window’s native maximize button state.
+   *
+   * #### Platform-specific
+   *
+   * - **Linux / iOS / Android:** Unsupported.
+   *
+   * @example
+   * ```typescript
+   * import { appWindow } from '@tauri-apps/api/window';
+   * const maximizable = await appWindow.isMaximizable();
+   * ```
+   *
+   * @returns Whether the window's native maximize button is enabled or not.
+   *  */
+  async isMaximizable(): Promise<boolean> {
+    return invokeTauriCommand({
+      __tauriModule: 'Window',
+      message: {
+        cmd: 'manage',
+        data: {
+          label: this.label,
+          cmd: {
+            type: 'isMaximizable'
+          }
+        }
+      }
+    })
+  }
+
+  /**
+   * Gets the window’s native minimize button state.
+   *
+   * #### Platform-specific
+   *
+   * - **Linux / iOS / Android:** Unsupported.
+   *
+   * @example
+   * ```typescript
+   * import { appWindow } from '@tauri-apps/api/window';
+   * const minimizable = await appWindow.isMinimizable();
+   * ```
+   *
+   * @returns Whether the window's native minimize button is enabled or not.
+   *  */
+  async isMinimizable(): Promise<boolean> {
+    return invokeTauriCommand({
+      __tauriModule: 'Window',
+      message: {
+        cmd: 'manage',
+        data: {
+          label: this.label,
+          cmd: {
+            type: 'isMinimizable'
+          }
+        }
+      }
+    })
+  }
+
+  /**
+   * Gets the window’s native close button state.
+   *
+   * #### Platform-specific
+   *
+   * - **Linux / iOS / Android:** Unsupported.
+   *
+   * @example
+   * ```typescript
+   * import { appWindow } from '@tauri-apps/api/window';
+   * const closable = await appWindow.isClosable();
+   * ```
+   *
+   * @returns Whether the window's native close button is enabled or not.
+   *  */
+  async isClosable(): Promise<boolean> {
+    return invokeTauriCommand({
+      __tauriModule: 'Window',
+      message: {
+        cmd: 'manage',
+        data: {
+          label: this.label,
+          cmd: {
+            type: 'isClosable'
           }
         }
       }
@@ -773,7 +921,6 @@ class WindowManager extends WebviewWindowHandle {
    * await appWindow.center();
    * ```
    *
-   * @param resizable
    * @returns A promise indicating the success or failure of the operation.
    */
   async center(): Promise<void> {
@@ -809,7 +956,7 @@ class WindowManager extends WebviewWindowHandle {
    * await appWindow.requestUserAttention();
    * ```
    *
-   * @param resizable
+   * @param requestType
    * @returns A promise indicating the success or failure of the operation.
    */
   async requestUserAttention(
@@ -859,6 +1006,105 @@ class WindowManager extends WebviewWindowHandle {
           cmd: {
             type: 'setResizable',
             payload: resizable
+          }
+        }
+      }
+    })
+  }
+
+  /**
+   * Sets whether the window's native maximize button is enabled or not.
+   * If resizable is set to false, this setting is ignored.
+   *
+   * #### Platform-specific
+   *
+   * - **macOS:** Disables the "zoom" button in the window titlebar, which is also used to enter fullscreen mode.
+   * - **Linux / iOS / Android:** Unsupported.
+   *
+   * @example
+   * ```typescript
+   * import { appWindow } from '@tauri-apps/api/window';
+   * await appWindow.setMaximizable(false);
+   * ```
+   *
+   * @param maximizable
+   * @returns A promise indicating the success or failure of the operation.
+   */
+  async setMaximizable(maximizable: boolean): Promise<void> {
+    return invokeTauriCommand({
+      __tauriModule: 'Window',
+      message: {
+        cmd: 'manage',
+        data: {
+          label: this.label,
+          cmd: {
+            type: 'setMaximizable',
+            payload: maximizable
+          }
+        }
+      }
+    })
+  }
+
+  /**
+   * Sets whether the window's native minimize button is enabled or not.
+   *
+   * #### Platform-specific
+   *
+   * - **Linux / iOS / Android:** Unsupported.
+   *
+   * @example
+   * ```typescript
+   * import { appWindow } from '@tauri-apps/api/window';
+   * await appWindow.setMinimizable(false);
+   * ```
+   *
+   * @param minimizable
+   * @returns A promise indicating the success or failure of the operation.
+   */
+  async setMinimizable(minimizable: boolean): Promise<void> {
+    return invokeTauriCommand({
+      __tauriModule: 'Window',
+      message: {
+        cmd: 'manage',
+        data: {
+          label: this.label,
+          cmd: {
+            type: 'setMinimizable',
+            payload: minimizable
+          }
+        }
+      }
+    })
+  }
+
+  /**
+   * Sets whether the window's native close button is enabled or not.
+   *
+   * #### Platform-specific
+   *
+   * - **Linux:** GTK+ will do its best to convince the window manager not to show a close button. Depending on the system, this function may not have any effect when called on a window that is already visible
+   * - **iOS / Android:** Unsupported.
+   *
+   * @example
+   * ```typescript
+   * import { appWindow } from '@tauri-apps/api/window';
+   * await appWindow.setClosable(false);
+   * ```
+   *
+   * @param closable
+   * @returns A promise indicating the success or failure of the operation.
+   */
+  async setClosable(closable: boolean): Promise<void> {
+    return invokeTauriCommand({
+      __tauriModule: 'Window',
+      message: {
+        cmd: 'manage',
+        data: {
+          label: this.label,
+          cmd: {
+            type: 'setClosable',
+            payload: closable
           }
         }
       }
@@ -2037,6 +2283,27 @@ class WebviewWindow extends WindowManager {
     }
     return null
   }
+
+  /**
+   *  Gets the focused window.
+   * @example
+   * ```typescript
+   * import { WebviewWindow } from '@tauri-apps/api/window';
+   * const focusedWindow = WebviewWindow.getFocusedWindow();
+   * ```
+   *
+   * @returns The WebviewWindow instance to communicate with the webview or `undefined` if there is not any focused window.
+   *
+   * @since 1.4
+   */
+  static async getFocusedWindow(): Promise<WebviewWindow | null> {
+    for (const w of getAll()) {
+      if (await w.isFocused()) {
+        return w
+      }
+    }
+    return null
+  }
 }
 
 /** The WebviewWindow for the current window. */
@@ -2152,6 +2419,18 @@ interface WindowOptions {
    * The user agent for the webview.
    */
   userAgent?: string
+  /**
+   * Whether the window's native maximize button is enabled or not. Defaults to `true`.
+   */
+  maximizable?: boolean
+  /**
+   * Whether the window's native minimize button is enabled or not. Defaults to `true`.
+   */
+  minimizable?: boolean
+  /**
+   * Whether the window's native close button is enabled or not. Defaults to `true`.
+   */
+  closable?: boolean
 }
 
 function mapMonitor(m: Monitor | null): Monitor | null {